L6 Devide functions to separate files
