+++ src/js/buildSizesInCard.js
@@ -1,0 +1,16 @@
+const buildSizesInCard = function (el) {
+    if (el.length > 0) {
+        let html = '<span style="text-transform: uppercase;">';
+
+        el.forEach((c, i) => {
+            html += `${i < 1 ? '' : '/'}${c}`;
+        });
+        html += '</span>';
+
+        return html;
+    }
+
+    return '&nbsp';
+};
+
+export { buildSizesInCard as default };
